Whitespace formatting
Try to follow best practices when coding.
For example indent lines by 4 spaces and not by 2 spaces or tabs.

Don't leave a space between the name of a function and the parentheses
containing its arguments or between the keyword template and its angle
brackets.

Also leave a space to the left and right of binary operators such as +,
-, *, /.