Reduce parameter list
Only pass the arguments absolutely necessary in your functions, i.e. the
min and max value needed for counting sort can be calculated inside the
function, so that the user of your library doesn't have to calculate
them themselves.

Also, the keyword 'end' is reserved, and its use as as a variable name
should be deferred.